refactor(tools): modularize server.py into tools/ directory (#28, #3)
Split monolithic server.py (57 tools) into per-entity modules under
tools/. Pure refactor with zero functional changes — all 57 tools
register with identical names, descriptions, and parameter signatures.
